working on initializing the game memory arena

+++ src/math/itfliesby-math.hpp
@@ -4,9 +4,9 @@
 #include "itfliesby-math-vec2.hpp"
 #include "itfliesby-math-mat2.hpp"
 
-#define ITFLIESBY_MATH_KILOBYTES(value) (value * 1024)
-#define ITFLIESBY_MATH_MEGABYTES(value) (ITFLIESBY_MATH_KILOBYTES(value) * 1024)
-#define ITFLIESBY_MATH_GIGABYTES(value) (ITFLIESBY_MATH_MEGABYTES(value) * 1024)
+#define ITFLIESBY_MATH_KILOBYTES(value) ((long long)(value * 1024))
+#define ITFLIESBY_MATH_MEGABYTES(value) ((long long)(ITFLIESBY_MATH_KILOBYTES(value) * 1024))
+#define ITFLIESBY_MATH_GIGABYTES(value) ((long long)(ITFLIESBY_MATH_MEGABYTES(value) * 1024))
